+++ RayTracer/src/Geometry/SphereGroup.h
@@ -24,7 +24,7 @@ public:
 	bool intersect(Ray& ray, HitInfo& hit, unsigned int i) const { return spheres[i].intersect(ray, hit, 0); }
 	
 	/// return the number of faces in the mesh
-	size_t num_primitives() const { return spheres.size(); }
+	unsigned int num_primitives() const { return static_cast<unsigned int>(spheres.size()); }
 
 	/// returns the bbox for the whole mesh
 	AABB get_bbox() const;
